+++ pkgbuilder/__init__.py
@@ -1,7 +1,7 @@
 # -*- encoding: utf-8 -*-
-# PKGBUILDer v4.2.17
+# PKGBUILDer v4.2.18
 # An AUR helper (and library) in Python 3.
-# Copyright © 2011-2018, Chris Warrick.
+# Copyright © 2011-2019, Chris Warrick.
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -35,7 +35,7 @@
 """
 An AUR helper (and library) in Python 3.
 
-:Copyright: © 2011-2018, Chris Warrick.
+:Copyright: © 2011-2019, Chris Warrick.
 :License: BSD (see /LICENSE).
 """
 
@@ -45,7 +45,7 @@ import sys
 
 
 __title__ = 'PKGBUILDer'
-__version__ = '4.2.17'
+__version__ = '4.2.18'
 __author__ = 'Chris Warrick'
 __license__ = '3-clause BSD'
 __docformat__ = 'restructuredtext en'