+++ docs/CHANGELOG.rst
@@ -3,10 +3,10 @@ Appendix C. Changelog
 =====================
 :Info: This is the changelog for PKGBUILDer.
 :Author: Chris Warrick <chris@chriswarrick.com>
-:Copyright: © 2011-2018, Chris Warrick.
+:Copyright: © 2011-2019, Chris Warrick.
 :License: BSD (see /LICENSE or :doc:`Appendix B <LICENSE>`.)
-:Date: 2018-03-25
-:Version: 4.2.17
+:Date: 2018-07-31
+:Version: 4.2.18
 
 .. index:: CHANGELOG
 
@@ -46,6 +46,9 @@ More information can be found on GitHub in the `releases section <https://github
 
 Version History
 ===============
+
+4.2.18
+    * Add a ``paccommand`` option to the INI file.
 
 4.2.17
     * Add ``-G``, ``--get`` as ``-F``, ``--fetch`` alternatives (Issue #56)